+++ controllers/contactControllers.js
@@ -1,16 +1,8 @@
-import Joi from 'joi';
 
 import contactsService from '../models/contacts.js';
 import { HttpError } from '../helpers/index.js';
 import {ctrlWrapper} from '../decorators/index.js';
 
-const contactsAddSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().email().required(),
-  phone: Joi.string()
-    .pattern(/^(\+\d{1,3})?[-()\d\s]+$/)
-    .required(),
-});
 
 const getAll = async (req, res) => {
     const result = await contactsService.listContacts();
@@ -20,45 +12,38 @@ const getAll = async (req, res) => {
 const getById = async (req, res) => {
     const { contactId } = req.params;
     const result = await contactsService.getContactById(contactId);
-
     if (!result) {
-      throw HttpError(404, 'Contact not found');
+      throw HttpError(404, 'Not found');
     }
     res.json(result);
 
 }
 
-const add = async (req, res) => {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
+const add = async (req, res) => { 
     const result = await contactsService.addContact(req.body);
     res.status(201).json(result);
-
 }
 
 const deleteById = async (req, res) => {
     const { contactId } = req.params;
     const result = await contactsService.removeContact(contactId);
     if (!result) {
-      throw HttpError(404, 'Contact not found');
+      throw HttpError(404, 'Not found');
     }
     res.json({ message: 'Contact deleted!' });
 }
 
 const updateById = async (req, res) => {
-    const { error } = contactsAddSchema.validate(req.body);
-    if (error) {
-      throw HttpError(400, error.message);
-    }
+  if (Object.keys(req.body).length === 0) {
+    throw HttpError(400, 'missing fields');
+  }
     const { contactId } = req.params;
     const result = await contactsService.updateContact(contactId, req.body);
     if (!result) {
-      throw HttpError(404, 'Contact not found');
+        throw HttpError(404, 'Not found');
     }
-    res.json(result);
 
+    res.json(result);
 }
 
 export default {
