+++ decorators/index.js
@@ -1,1 +1,2 @@
 export { default as ctrlWrapper } from './ctrlWrapper.js';
+export { default as validateBody } from './validateBody.js';
