+++ models/contacts.json
@@ -1,15 +1,9 @@
 [
   {
-    "id": "1DEXoP8AuCGYc1YgoQ6hw",
-    "name": "Abbot Franks",
-    "email": "scelerisque@magnis.org",
-    "phone": "(186) 568-3720"
-  },
-  {
     "id": "Z5sbDlS7pCzNsnAHLtDJd",
-    "name": "Reuben Henry",
+    "name": "Chack Noris",
     "email": "pharetra.ut@dictum.co.uk",
-    "phone": "(715) 598-5792"
+    "phone": "(003) 002-666"
   },
   {
     "id": "C9sjBfCo4UJCWjzBnOtxl",
@@ -37,6 +31,18 @@
   },
   {
     "id": "rIMCBeYLJjscN-2kicOKq",
+    "name": "Chack Noris",
+    "email": "punch.in@egetlacus.ca",
+    "phone": "(003) 002-666"
+  },
+  {
+    "id": "xV9Y3ZSLtBpfrAN3T7fU_",
+    "name": "Chack Noris",
+    "email": "punch.in@egetlacus.ca",
+    "phone": "(003) 002-666"
+  },
+  {
+    "id": "RBAGcIMvJM9SCnVPRELZQ",
     "name": "Chack Noris",
     "email": "punch.in@egetlacus.ca",
     "phone": "(003) 002-666"
