+++ routes/api/contacts.js
@@ -2,16 +2,22 @@ import express from 'express';
 
 import contactsController from '../../controllers/contactControllers.js'
 
+import contactsSchema from '../../schema/schemaJoi.js'
+
+import { validateBody } from '../../decorators/index.js';
+
+import {isEmptyBody} from '../../middlewars/index.js';
+
 const contactsRouter = express.Router();
 
 contactsRouter.get('/', contactsController.getAll);
 
 contactsRouter.get('/:contactId', contactsController.getById);
 
-contactsRouter.post('/', contactsController.add);
+contactsRouter.post('/',validateBody(contactsSchema.contactsAddSchema), contactsController.add);
 
 contactsRouter.delete('/:contactId', contactsController.deleteById);
 
-contactsRouter.put('/:contactId', contactsController.updateById);
+contactsRouter.put('/:contactId', isEmptyBody, contactsController.updateById);
 
 export default contactsRouter;
